requirements.txt
pandas
numpy
plotly
python-dotenv
optuna
datasets
transformers[torch]
pillow
nltk
torch torchvision # for GPU use
wandb
evaluate
rouge-score
absl-py
h5py
# ray # whl must be downloaded and installed for windows and python version